Pre Alpha 1.1.18
Added an item holding mechanism to the player that allows for any prop to be held as long as its origin is in the center. The recommended size for the prop is 10x10 to make it seem realistic. That way we do not have to make separate Fritz sprites for each prop.

The mechanism works as long as it knows what sprite to use, so I also added two variable called equipped_object (which holds a sprite index of the object, otherwise "noone" by default), and is_holding_item which is a bool toggled by the hold item key to make Fritz put the item in his mouth (for now its E, but we'll change it later)

equipped_object will change depending on what the player equips from the ground or picks from inventory.